Create Shortest Uncommon Subsequence
https://practice.geeksforgeeks.org/problems/shortest-uncommon-subsequence/0

Given two strings S and T, find length of the shortest subsequence in S which is not a subsequence in T. If no such subsequence is possible, return -1. A subsequence is a sequence that appears in the same relative order, but not necessarily contiguous. A string of length n has 2^n different possible subsequences.

 

Input:

The first line of input contains an integer T denoting the number of test cases. Then T test cases follow. The first line of each test case contains integers M and N where M denotes the length of first string and N denotes the length of second string.

The second line of each test case contains the two string S and T. Both string are comprised only of lower case alphabets.


Output:

Output the length of the shortest uncommon subsequence for each test case in a new line. If no such sub sequence is possible print -1.