Create Maximum Rectangular Area in Histogram
Idea is to calculate area with all the bars of the histogram as the height of rectangle one by one and return the maximum. 
For that, we maintain a stack of indices, where we keep on pushing the larger bars. On coming through a smaller bar as the top of the stack, we know that the right limit of the present top has been met. Also, the left limit is the element just below the top. Hence area for the particular bar can be calculated.